feat: 단위 테스트를 위해 s3Url을 만드는 클래스 추가
