+++ src/main/java/com/map/gaja/client/infrastructure/image/S3UrlGenerator.java
@@ -1,0 +1,25 @@
+package com.map.gaja.client.infrastructure.image;
+
+import org.springframework.beans.factory.annotation.Value;
+
+import javax.annotation.PostConstruct;
+
+public class S3UrlGenerator {
+    @Value("${cloud.aws.s3.bucket}")
+    private String bucket;
+
+    @Value("${cloud.aws.region.static}")
+    private String region;
+
+    private String s3Url;
+
+    @PostConstruct
+    private void init() {
+        String s3UrlFormat = "https://%s.s3.%s.amazonaws.com/";
+        s3Url = String.format(s3UrlFormat, bucket, region);
+    }
+
+    public String getS3Url() {
+        return s3Url;
+    }
+}
